cmake: more visibility for NO_GETENV
Majority of libjpeg-turbo users are totally unaware of the built-in ability
to control memory usage and/or code paths used by library by means of
environment variables. This might be desireable for performance testing
purposes, but in real life it's only dangerous (library will crash when forced
to use particular code paths on CPUs that don't support them).

NO_GETENV define is already in library code, the only thing this patch does
is adding it to CMake so it can be seen without digging through the code.

Signed-off-by: Piotr Dałek <git@predictor.org.pl>