hacky patch to make the three bit depths compile (8, 12 and 16 bits). Still there's *some* trouble to get it built...
TODO/NOTE: the 12-bits and 16-bits directories are the quick hack to allow MSVC2022 to build all jpeg modes in monolithic build mode. HOWEVER, the list of source files theere needs to be pruned as it was a simple script (mk12.sh) that did this for us and we don't need to xref them all like that!